requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.4.0
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-import-export==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
drf-api-tracking==1.6.0
et-xmlfile==1.0.1
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.17.0
joblib==0.14.1
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupPy==1.14
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.0
redis==3.5.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
vine==1.3.0
wcwidth==0.1.9
wrapt==1.12.1
xgboost==1.0.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
